Move BTPayPalError into its own file
