+++ Sources/BraintreePayPal/BTPayPalClientSwift.swift
@@ -2,24 +2,6 @@ import Foundation
 import AuthenticationServices
 import BraintreeCore
 
-/// Error codes associated with PayPal.
-public enum BTPayPalError: Int {
-    /// Unknown error
-    case unknown
-    
-    /// PayPal is disabled in configuration
-    case disabled
-    
-    /// Invalid request, e.g. missing PayPal request
-    case invalidRequest
-    
-    /// Braintree SDK is integrated incorrectly
-    case integration
-    
-    /// Payment flow was canceled, typically initiated by the user when exiting early from the flow.
-    case canceled
-}
-
 @objcMembers public class BTPayPalClientSwift: NSObject {
     
     // MARK: - Internal Properties
